final-project/python-test/requirements.txt
websocket-client
rel
python-dotenv
